requirements.txt
cloudpickle==3.0.0
numpy==1.26.4
pandas==2.2.0
pyjanitor==0.23.1
scikit_learn==1.2.2
seaborn==0.13.2
streamlit==1.31.0
xgboost==2.0.3
